program.
A straightforward yet powerful C program called mathematicsTOOL01 can identify if a given integer is prime or composite. The user enters a number, and the application quickly determines whether or not it can be divided by whole numbers. A number is considered prime if it can be divided only by 1 and itself. If not, it is categorized as Composite. Primality exams can be completed fast using this program, and it can help you understand the fundamentals of number theory.